% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{addAssay}
\alias{addAssay}
\title{Assay setter}
\usage{
addAssay(segmentation, assay, name)
}
\arguments{
\item{segmentation}{a \code{segmentation} class object}

\item{assay}{a \code{segmentationAssay} class object}

\item{name}{character, the name of the assay}
}
\description{
Function to set the \code{assays} slot in \code{segmentation} class objects
}
