% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{assays,segmentation-method}
\alias{assays,segmentation-method}
\title{Assay getter (method)}
\usage{
\S4method{assays}{segmentation}(x)
}
\arguments{
\item{x}{a \code{segmentation} class object}
}
\description{
S4 method to get the \code{assays} slot in \code{segmentation} class objects
}
