% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{box_make}
\alias{box_make}
\title{Create a bounding box}
\usage{
box_make(point_set, step_size = 1)
}
\arguments{
\item{point_set}{a data frame containing point coordinates. Columns must be named as \code{"x"} and \code{"y"}.}

\item{step_size}{a numeric indicating how much bigger the bounding box should be compared to the point cloud. Default is 1 unit.}
}
\value{
A data frame containing the original points and the bounding box coordinates
}
\description{
Generates a rectangular bounding box around a point cloud, filling it with 0 wherever there are no points
}
