% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldcuts.R
\name{dir_change}
\alias{dir_change}
\title{Change array directions}
\usage{
dir_change(array, from = "PIR", to = "RAS")
}
\arguments{
\item{array}{a 3D array containing voxel data.}

\item{from}{character, a 3-word direction code that the array is in. One of \code{PIR}, \code{RAS}, or \code{LPS}.}

\item{to}{character, a 3-word direction code that the array will be rotated to. One of \code{PIR}, \code{RAS}, or \code{LPS}.}
}
\value{
a 3D array rotated according to original and destination directions
}
\description{
Rotates the array so that it goes from one orientation to another. Currently supports PIR, RAS and LPS. Depends on \code{freesurferformats}.
}
