% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{drawMaxProjection}
\alias{drawMaxProjection}
\title{Create a maximum projection}
\usage{
drawMaxProjection(
  str,
  polygon_list,
  subset_slices = NULL,
  make_polygon = FALSE
)
}
\arguments{
\item{str}{a character containing a structure id}

\item{polygon_list}{a polygon set (list) in which structures should be found}

\item{subset_slices}{a numeric vector indicating which slices should be used to draw the projection. Default is \code{NULL}, using all slices}

\item{make_polygon}{logical, should the resulting polygon be created? Default is \code{FALSE}}
}
\value{
a data frame containing coordinates of points of the maximum projection, or the ordered coordinates of a(the) polygon(s) enclosing all points.
}
\description{
Creates a point set or a polygon (ordered set of xy coordinates) resulting as the projection on the same plane of polygons in all slices.
}
