% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldcuts.R
\name{drawOutline}
\alias{drawOutline}
\title{Draw maximum outline}
\usage{
drawOutline(M, plane)
}
\arguments{
\item{M}{a molten array (of class `data.table`) containing coordinates}
}
\value{
a data frame containing ordered coordinates for the outline polygon
}
\description{
Draws a polygon obtained by projecting all polygons on the plane
}
