% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldcuts.R
\name{dummy_cubes}
\alias{dummy_cubes}
\title{Build dummy cubes}
\usage{
dummy_cubes(n)
}
\arguments{
\item{n}{the number of voxels in the side of each cube}
}
\value{
a 3D array containing 8 different cubes valued 1 to 8
}
\description{
Builds a 3D array with 8 dummy cubes, for demo/testing purposes only.
}
