% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldcuts.R
\name{dummy_ontology}
\alias{dummy_ontology}
\title{Build dummy ontology}
\usage{
dummy_ontology()
}
\value{
an ontology table for 8 cubes named A to H.
}
\description{
Builds an ontology table to go with the output of \code{dummy_cubes()}. For demo/testing purposes only.
}
