% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{find_holes}
\alias{find_holes}
\title{Find holes}
\usage{
find_holes(ib_df, which_fun = any)
}
\arguments{
\item{ib_df}{a data.frame containing isobands}

\item{which_fun}{function to determine how to count inclusion, default is \code{any()}}
}
\value{
a data frame where polygon clusters are labelled as holes or not
}
\description{
Function to find holes in a set of polygons
}
