% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxprojection.R
\name{getCenter}
\alias{getCenter}
\title{Get visual center}
\usage{
getCenter(poly)
}
\arguments{
\item{poly}{a polygon in array or matrix form, in which the first two columns contain x and y vertex coordinates}
}
\value{
a vector containing the coordinates of the visual center of the input polygon
}
\description{
Finds the visual center (pole of inaccessibility) for a polygon.
}
