% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldcuts.R
\name{getSlice}
\alias{getSlice}
\title{Get a specific slice from a segmentation}
\usage{
getSlice(segmentation, plane, slice, fill = FALSE)
}
\arguments{
\item{segmentation}{a \code{segmentation} class object.}

\item{plane}{character, name of the plane (one of `sagittal`, `coronal`, or `axial`)}

\item{slice}{numeric or character, the slice number. If numeric, it will be used as index of the slice. If character, it will be the slice as named in the segmentation.}

\item{fill}{logical, should the function return just the polygon vertices or all voxels? Default is \code{FALSE}.}
}
\value{
a `data.frame` containing the x y coordinates for all points within the slice (filled polygons)
}
\description{
Renders a slice from a plane of a segmentation as a data frame
}
