% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxprojection.R
\name{makeOntologyGraph}
\alias{makeOntologyGraph}
\title{Plot the ontology graph}
\usage{
makeOntologyGraph(o, circular = TRUE)
}
\arguments{
\item{o}{a data frame containing the ontology/tree}

\item{circular}{a logical indicating whether the layout should be circular}
}
\value{
a `ggplot` object with the graph
}
\description{
Creates a `ggplot` plot of the ontology graph/tree
}
