% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{makeOntologyPalettes}
\alias{makeOntologyPalettes}
\title{Create a palette for an ontology/tree}
\usage{
makeOntologyPalettes(
  o,
  graph_depth = NULL,
  start_nodes = NULL,
  colorspace = "pretty",
  default_col = "white"
)
}
\arguments{
\item{o}{a data frame containing the ontology/tree}

\item{graph_depth}{a numeric defining the graph depth at which the initial colouring should be resolved. Can only be used if \code{start_nodes} is NULL}

\item{start_nodes}{a character vector with the acronyms of the nodes that are being selected for initial colouring}

\item{colorspace}{the color space to be used as input for \code{qualpal} - see \code{?qualpal} for more information. Defaults to \code{"pretty"}.}

\item{default_col}{a character indicating the color to be used for nodes that will not be colored}
}
\value{
a vector of colors ordered by the same order as the nodes
}
\description{
Generates a maximally separated color palette for tree nodes taking into account the level at which colors should be initially resolved
}
