% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxprojection.R
\name{makeSegmentation}
\alias{makeSegmentation}
\title{Make segmentation}
\usage{
makeSegmentation(
  nifti_file = NULL,
  ontology_file,
  array = NULL,
  molten_array = NULL,
  outliers = NULL,
  verbose = TRUE,
  directions_from = "RAS",
  directions_to = "RAS",
  planes = "all",
  draw_outline = TRUE,
  subset_sagittal = NULL,
  subset_coronal = NULL,
  subset_axial = NULL,
  parallel = TRUE
)
}
\arguments{
\item{nifti_file}{a character string pointing to the NIfTI file address}

\item{ontology_file}{a character string pointing to the ontology .csv file address}

\item{array}{a 3D array containing structure annotations (optional if \code{nifti_file} and \code{molten_array} are not specified). Default is \code{NULL}}

\item{molten_array}{an array in molten form, such as the result of \code{reshape2::melt()} (optional if \code{nifti_file} and \code{array} are not specified). Default is \code{NULL}}

\item{outliers}{a numeric vector indicating outlier points to be eliminated before drawing polygons. Default is \code{NULL}}

\item{verbose}{logical, should messages on the process be displayed? Default is \code{TRUE}}

\item{directions_from}{character, indicates the original orientation of the array. Default is \code{"RAS"} (Left to Right, Posterior to Anterior, Inferior to Superior).}

\item{directions_to}{character, indicates the final orientation of the array. The array is rotated only if it is different from \code{directions_from}. Default is \code{"RAS"}, which results in the proper assignment of axis labels.}

\item{planes}{character, either \code{"all"} (default) or any subset of \code{"sagittal"}, \code{"coronal"}, and \code{"axial"}. Determines along which axes the array wil be sliced.}

\item{draw_outline}{logical, should the outline of the whole array be drawn? Default is `TRUE`}

\item{subset_sagittal}{numeric vector, indicates a subset of slices along the sagittal plane that will be drawn. Default is \code{NULL}}

\item{subset_coronal}{numeric vector, indicates a subset of slices along the coronal plane that will be drawn. Default is \code{NULL}}

\item{subset_axial}{numeric vector, indicates a subset of slices along the axial plane that will be drawn. Default is \code{NULL}}

\item{parallel}{logical, should the polygons be drawn using parallel processing? Uses \code{future.apply::future_lapply()}, so the proper setup must be done beforehand. Default is \code{FALSE}.}
}
\value{
a `segmentation` S4 object containing the following slots:
}
\description{
Creates a `segmentation` object from an array or a `NIfTI` file
}
