% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{metaData,segmentation-method}
\alias{metaData,segmentation-method}
\title{Metadata getter (method)}
\usage{
\S4method{metaData}{segmentation}(x)
}
\arguments{
\item{x}{a \code{segmentation} class object}
}
\description{
S4 method to access the \code{metadata} slot in \code{segmentation} class objects
}
