% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{ontology,segmentation-method}
\alias{ontology,segmentation-method}
\title{Ontology getter (method)}
\usage{
\S4method{ontology}{segmentation}(x)
}
\arguments{
\item{x}{a \code{segmentation} class object}
}
\description{
S4 method to get the \code{ontology} slot in \code{segmentation} class objects
}
