% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{ontology<-,segmentation-method}
\alias{ontology<-,segmentation-method}
\title{Ontology setter (method)}
\usage{
\S4method{ontology}{segmentation}(x) <- value
}
\arguments{
\item{x}{a \code{segmentation} class object}

\item{value}{a table containing the ontology tree}
}
\description{
S4 method to set the \code{ontology} slot in \code{segmentation} class objects, with sanity checks.
}
