% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{ontology<-}
\alias{ontology<-}
\title{Ontology setter (generic)}
\usage{
ontology(x) <- value
}
\arguments{
\item{x}{a \code{segmentation} class object}

\item{value}{a table containing the ontology tree}
}
\description{
S4 method to set the \code{ontology} slot in \code{segmentation} class objects
}
