% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{ontology}
\alias{ontology}
\title{Ontology getter (generic)}
\usage{
ontology(x)
}
\arguments{
\item{x}{a \code{segmentation} class object}
}
\description{
S4 method to get the \code{ontology} slot in \code{segmentation} class objects
}
