% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldcuts.R
\name{outline_draw}
\alias{outline_draw}
\title{Draw maximum outline}
\usage{
outline_draw(M, plane)
}
\arguments{
\item{M}{a molten array (of class `data.table`) containing coordinates}

\item{plane}{character, one of "sagittal", "coronal", "axial"}
}
\value{
a data frame containing ordered coordinates for the outline polygon
}
\description{
Draws a polygon obtained by projecting all polygons on the plane
}
