% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldcuts.R
\name{plotBrainMap}
\alias{plotBrainMap}
\title{Plot numeric data on a segmentation}
\usage{
plotBrainMap(
  segmentation,
  feature,
  assay,
  plane = "sagittal",
  smooth = TRUE,
  smoothness = 3,
  min_points = 5,
  show_labels = TRUE
)
}
\arguments{
\item{segmentation}{a \code{segmentation} class object.}

\item{feature}{character, feature to be plotted from the \code{assay} slot.}

\item{assay}{character, the name of the \code{assay} slot in the \code{segmentation}.}

\item{plane}{character, the plane for the maximum projection. Default is "sagittal".}

\item{smooth}{logical, should shapes be smoothed? Default is \code{TRUE}.}

\item{smoothness}{numeric, the smoothing to be used. Default is 3.}

\item{min_points}{numeric, the minimum points to smooth polygons. Default is 5.}

\item{max_projection}{logical, should the maximum projection be computed beforehand? Default is \code{TRUE}.}
}
\value{
a `ggplot` plot in which structures are coloured according to a numeric value
}
\description{
Creates a `ggplot` plot where numeric data, mapped to structures, fill the segmentation shapes in a colour scale
}
