% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldcuts.R
\name{plotMaxProjection}
\alias{plotMaxProjection}
\title{Plot a projection from a segmentation}
\usage{
plotMaxProjection(
  segmentation,
  plane,
  name,
  smooth = TRUE,
  smoothness = 3,
  show_labels = FALSE
)
}
\arguments{
\item{segmentation}{a \code{segmentation} class object.}

\item{plane}{character, name of the plane (one of `sagittal`, `coronal`, or `axial`)}

\item{name}{character, name of the projection to be plotted from the \code{segmentation} slot.}

\item{smooth}{logical, should polygons be smoothed? Default is \code{TRUE}}

\item{smoothness}{numeric, the kernel bandwidth for kernel smoothing. Default is 3.}

\item{show_labels}{logical, should structure acronyms be shown as labels? Default is \code{FALSE}}
}
\value{
a `ggplot` plot of the projection in both directions
}
\description{
Plots a named projection from a segmentation using the ontology color scheme
}
