% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldcuts.R
\name{plotOntologyGraph}
\alias{plotOntologyGraph}
\title{Plot the ontology graph}
\usage{
plotOntologyGraph(seg, circular = TRUE)
}
\arguments{
\item{circular}{a logical indicating whether the layout should be circular}

\item{o}{a data frame containing the ontology/tree}
}
\value{
a `ggplot` object with the graph
}
\description{
Creates a `ggplot` plot of the ontology graph/tree
}
