% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{poly_build}
\alias{poly_build}
\title{Build a polygon}
\usage{
poly_build(seg_point_set)
}
\arguments{
\item{seg_point_set}{A  \code{segPointSet} object}
}
\value{
A data frame containing polygon coordinates, \code{structure}, \code{id}, and \code{subid}
}
\description{
Creates a polygon in a data frame from a \code{segPointSet} object
}
