% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{poly_fill}
\alias{poly_fill}
\title{Fill a polygon with points}
\usage{
poly_fill(point_set, step_size = 1)
}
\arguments{
\item{point_set}{a polygon point set in array or matrix form, in which the first two columns contain x and y vertex coordinates.}

\item{step_size}{a numeric indicating the spacing of the points}
}
\value{
A data frame containing coordinates for points within the polygon, including polygon vertices, inheriting all the other columns from the input polygon point set.
}
\description{
Fills a polygon with regularly spaced points.
}
