% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{poly_make}
\alias{poly_make}
\title{Make polygon}
\usage{
poly_make(point_set, step_size = 1)
}
\arguments{
\item{point_set}{a polygon in array or matrix form, in which the first two columns contain x and y vertex coordinates.}

\item{step_size}{a numeric indicating the spacing of the points}
}
\value{
A data frame containing coordinates for polygon vertices and the slice it originated from (if present in the original point set).
}
\description{
Makes a polygon from a point set
}
