% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{projections,segmentation-method}
\alias{projections,segmentation-method}
\title{Projection getter (method)}
\usage{
\S4method{projections}{segmentation}(x, name)
}
\arguments{
\item{x}{a \code{segmentation} class object}

\item{name}{character, the name of the projection}
}
\description{
S4 method to get the \code{projections} slot in \code{segmentation} class objects
}
