% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{projections}
\alias{projections}
\title{Projection getter (generic)}
\usage{
projections(x, name)
}
\arguments{
\item{x}{a \code{segmentation} class object}

\item{name}{character, the name of the projection}
}
\description{
S4 method to get the \code{projections} slot in \code{segmentation} class objects
}
