% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldcuts3D.R
\name{seg_build_meshes}
\alias{seg_build_meshes}
\title{Build segmentation meshes in 3D}
\usage{
seg_build_meshes(
  segmentation,
  subset_str = NULL,
  pct_reduce = 0.1,
  verbose = FALSE
)
}
\arguments{
\item{segmentation}{a `segmentation` class object}

\item{subset_str}{a character vector indicating the structure(s) to be rendered. Default is \code{NULL}, meaning the whole segmentation will be rendered as a single mesh.}

\item{pct_reduce}{numeric, target percentage of vertices for downsampling and remeshing. Closer to 0 gives a simpler mesh. Default is 0.1.}

\item{verbose}{logical, show progress of the meshing? Default is \code{FALSE}.}
}
\value{
a smoothed 3d mesh (as `mesh3d` object) from one or more structures
}
\description{
Uses marching cubes and mesh reduction to render the segmentation in 3 dimensions
}
