% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{seg_citation_add}
\alias{seg_citation_add}
\title{Citation setter}
\usage{
seg_citation_add(segmentation, name, citation)
}
\arguments{
\item{segmentation}{a \code{segmentation} class object}

\item{name}{character, the name of the citation (e.g. "Lebowski2001")}

\item{citation}{character, the citation}
}
\description{
Function to set the \code{citations} item in the metadata of \code{segmentation} class objects
}
