% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{seg_feature_plot}
\alias{seg_feature_plot}
\title{Plot numeric data on a segmentation}
\usage{
seg_feature_plot(
  segmentation,
  feature,
  assay,
  projection = NULL,
  plane = "sagittal",
  by = NULL,
  aggr_fun = NULL,
  rng = NULL,
  smooth = TRUE,
  smoothness = 3,
  minsize = 10,
  color_pal = NULL,
  show_side = "both",
  show_labels = TRUE,
  labelsize = 2,
  remove_axes = TRUE
)
}
\arguments{
\item{segmentation}{a \code{segmentation} class object.}

\item{feature}{character, feature to be plotted from the \code{assay} slot.}

\item{assay}{character, the name of the \code{assay} slot in the \code{segmentation}.}

\item{projection}{Character, name of the maximum projection slot to be used? Default is \code{NULL}, which uses the same name as selected in \code{assay}.}

\item{plane}{character, the plane for the maximum projection. Default is "sagittal".}

\item{by}{character with two elements, [1] the name of the column in the \code{sampledata} of the \code{assay} to be used for subsetting and aggregating and [2] the value to which the column in [1] should be equal to}

\item{aggr_fun}{function, the function to be used to aggregate samples across rows when using \code{by}}

\item{rng}{numeric with two elements with lower and upper bound for the color scale. Default is \code{NULL}}

\item{smooth}{logical, should shapes be smoothed? Default is \code{TRUE}.}

\item{smoothness}{numeric, the smoothing to be used. Default is 3.}

\item{minsize}{numeric, minimum number of vertices to draw a polygon. Default is 10.}

\item{color_pal}{character, the color palette to be used. The default is the `Sunset` palette from \code{colorspace}}

\item{show_side}{character, the side of the projection to be plotted. one of "first" ,"second", or "both". Default is "both".}

\item{show_labels}{logical, should segmentation labels be shown? Default is \code{TRUE}}

\item{labelsize}{numeric, the size of the labels. Default is 2.}

\item{remove_axes}{logical, should axes be shown? Default is \code{FALSE}}
}
\value{
a `ggplot` plot in which structures are coloured according to a numeric value
}
\description{
Creates a `ggplot` plot where numeric data, mapped to structures, fill the segmentation shapes in a colour scale
}
