% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldcuts3D.R
\name{seg_meshlist_add}
\alias{seg_meshlist_add}
\title{Add mesh list to a segmentation}
\usage{
seg_meshlist_add(
  segmentation,
  structures = NULL,
  pct_reduce = 0.1,
  puffiness = 1,
  verbose = FALSE
)
}
\arguments{
\item{segmentation}{a `segmentation` class object}

\item{structures}{a character vector of structure acronyms, indicating the structure(s) to be rendered. Default is \code{NULL}, meaning all structures will be rendered.}

\item{pct_reduce}{numeric, target percentage of vertices for downsampling and remeshing. Closer to 0 gives a simpler mesh. Default is 0.1.}

\item{puffiness}{numeric, determines a slight inflation of the mesh. Default is 1.}

\item{verbose}{logical, show progress of the meshing? Default is \code{FALSE}.}
}
\value{
a `segmentation` class object with a list of triangular meshes (as `mesh3d` objects) in the `meshes` slot.
}
\description{
Builds a list of structure meshes and stores it in the meshes slot.
}
