% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{seg_metadata<-,segmentation-method}
\alias{seg_metadata<-,segmentation-method}
\title{Metadata setter (method)}
\usage{
\S4method{seg_metadata}{segmentation}(x) <- value
}
\arguments{
\item{x}{a \code{segmentation} class object}

\item{value}{a list containing segmentation metadata}
}
\description{
S4 method to set the \code{metadata} slot in \code{segmentation} class objects
}
