% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{seg_plot}
\alias{seg_plot}
\title{Plot slices from a segmentation}
\usage{
seg_plot(
  segmentation,
  s_slice = NULL,
  c_slice = NULL,
  a_slice = NULL,
  smooth = TRUE,
  smoothness = 3,
  show_axis_rulers = TRUE,
  show_outline = TRUE,
  show_labels = FALSE,
  label_size = 2,
  minsize = 10,
  wrap_options = 1
)
}
\arguments{
\item{segmentation}{a \code{segmentation} class object}

\item{s_slice}{numeric, the slice number (index) on the sagittal plane}

\item{c_slice}{numeric, the slice number (index) on the coronal plane}

\item{a_slice}{numeric, the slice number (index) on the axial plane}

\item{smooth}{logical, should polygons be smoothed before drawing? Default is \code{TRUE}. Uses kernel smoothing.}

\item{smoothness}{numeric, the argument passed to the kernel smoothing function.}

\item{show_axis_rulers}{logical, should rulers for each axis be shown in each view? Default is \code{TRUE}.}

\item{show_outline}{logical, should the maximum segmentation outline be plotted? Default is \code{TRUE}.}

\item{show_labels}{logical, should structure acronym labels be shown on the plot? Default is \code{FALSE}.}

\item{label_size}{numeric, the size of the labels.}

\item{minsize}{numeric, the minimum number of points to create a polygon. Any structure in the slice with fewer points than this number will be discarded. Default is 10.}

\item{wrap_options}{numeric, the number of rows for the facet wrapping. Default is 1.}
}
\value{
a `ggplot` object with the graph
}
\description{
Creates a `ggplot` plot of up to 3 different slices (one per anatomical plane) of the segmentation
}
\examples{
array_cubes <- dummy_cubes(10)
ontology <- dummy_ontology()

seg <- seg_draw(
  array = array_cubes,
  ontology = ontology,
  verbose = FALSE
)

seg_plot(seg, s_slice = 10, a_slice = 4, c_slice = 2, smooth = FALSE)

# With labels
seg_plot(seg, s_slice = 10, a_slice = 4, c_slice = 2, smooth = FALSE, 
         show_labels = TRUE)
         
# Different wrap options:
seg_plot(seg, s_slice = 10, a_slice = 4, c_slice = 2, smooth = FALSE,
wrap_options = 2)         
 
}
