% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{seg_projection_add}
\alias{seg_projection_add}
\title{Create a maximum projection}
\usage{
seg_projection_add(
  name,
  structures,
  segmentation,
  planes_chosen = NULL,
  make_polygon = TRUE
)
}
\arguments{
\item{name}{a character containing the name of the maximum projection slot}

\item{structures}{a character containing one or more structure IDs}

\item{segmentation}{a \code{segmentation} class object}

\item{planes_chosen}{a character vector indicating which planes should be used to draw the projection. Default is \code{NULL}, using all planes}

\item{make_polygon}{logical, should the resulting polygon be created? Default is \code{TRUE}}
}
\value{
a \code{segmentation} object containing a maximum projection slot as a \code{segPointSet} object, with 2 views per selected plane.
}
\description{
Creates a point set or a polygon (ordered set of xy coordinates) resulting as the projection on the same plane of polygons in all slices.
}
