% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{seg_projection_remove}
\alias{seg_projection_remove}
\title{Remove a projection from a segmentation}
\usage{
seg_projection_remove(segmentation, name)
}
\arguments{
\item{segmentation}{a \code{segmentation} class object.}

\item{name}{character, name of the projection to be removed from the \code{segmentation} slot.}
}
\value{
a segmentation object without the named projection
}
\description{
Removes a named projection from a segmentation
}
