% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldcuts.R
\name{seg_select_str}
\alias{seg_select_str}
\title{Get a set of indices according to the ontology}
\usage{
seg_select_str(segmentation, structures, group_first = TRUE)
}
\arguments{
\item{segmentation}{a \code{segmentation} class object.}

\item{structures}{character, vector containing one or more structure acronyms}

\item{group_first}{logical, should priority be given to the group of structures or to the structure with the same acronym? Default is TRUE (prioritizing the group)}
}
\value{
a vector containing the indices of the structures
}
\description{
Retrieves all sets of indices belonging to a particular acronym
}
