% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{segmentationAssay-class}
\alias{segmentationAssay-class}
\title{S4 class representing an assay associated with a segmentation}
\description{
Holds numerical values for features in samples, and provides a one-to-many mapping for every sample to one or more structures
}
\section{Slots}{

\describe{
\item{\code{values}}{A `matrix` containing features (rows) in each sample (columns)}

\item{\code{mapping}}{A `list` in which each sample (list item name) contains the segmentation structural IDs to which samples are mapped}

\item{\code{sampledata}}{A `data.frame` containing other information on samples (e.g. description, sample metadata)}
}}

