% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{segPointSet-class}
\alias{segPointSet-class}
\title{S4 class representing a segmentation point set}
\description{
Holds all the data necessary to build a set of polygons via `buildPolygon` in order to be plotted.
}
\section{Slots}{

\describe{
\item{\code{coords}}{A `data.frame` containing ordered `x` and `y` coordinates for polygon drawing}

\item{\code{slice}}{A `character` containing the slice number}

\item{\code{structure}}{A `character` containing the structure ID}

\item{\code{id}}{A `character` containing the polygon ID (group level aesthetic for plotting)}

\item{\code{subid}}{A `character` containing the polygon sub-ID (subgroup level aesthetic for plotting)}
}}

