% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{show,segmentation-method}
\alias{show,segmentation-method}
\title{Show segmentation}
\usage{
\S4method{show}{segmentation}(object)
}
\arguments{
\item{object}{a \code{segmentation} class object}
}
\description{
S4 method to show the preview of a \code{segmentation} class objects
}
