% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{showCitation,segmentation-method}
\alias{showCitation,segmentation-method}
\title{Show citation (method)}
\usage{
\S4method{showCitation}{segmentation}(x)
}
\arguments{
\item{x}{a \code{segmentation} class object}
}
\description{
S4 method to show the \code{citation} slot of a \code{segmentation} class objects
}
