% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{showCitation}
\alias{showCitation}
\title{Show citation (generic)}
\usage{
showCitation(x)
}
\arguments{
\item{x}{a \code{segmentation} class object}
}
\description{
S4 method to show the \code{citation} slot of a \code{segmentation} class objects
}
