% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{slice_make}
\alias{slice_make}
\title{Make slices}
\usage{
slice_make(M, plane = "sagittal")
}
\arguments{
\item{M}{a data frame molten from a 3D array in long format, with x, y and z columns as \code{"Var1"}, \code{"Var2"}, \code{"Var3"} and the value column containing the voxel annotation}

\item{plane}{character, one of \code{"sagittal"}, \code{"coronal"}, or \code{"axial"}. The plane along which to slice the array. Default is \code{"sagittal"}}
}
\value{
A list of lists of structures, one list per slice
}
\description{
Divides a molten 3D array in slices along a selected plane
}
