% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxprojection.R
\name{sliceCheck}
\alias{sliceCheck}
\title{Check slices}
\usage{
sliceCheck(str, polygon_list)
}
\arguments{
\item{str}{a character containing a structure id}

\item{polygon_list}{a list of polygon sets generated by `makePolygonSets`}
}
\value{
a vector of indices of the polygon set `polygon_list` containing structure `str`
}
\description{
Checks which slices in a polygon list contain a given structure id
}
