% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{subsetByStructures}
\alias{subsetByStructures}
\title{Subset by structures}
\usage{
subsetByStructures(segmentation, structures, planes_chosen = NULL)
}
\arguments{
\item{segmentation}{an object of class \code{segmentation}}

\item{structures}{a vector of character structure IDs}

\item{planes_chosen}{a vector of planes to be further subset. Defaults to \code{NULL}, meaning the 3 anatomical planes will be considered.}
}
\value{
A \code{segmentation} class object containing only the structures defined in \code{structures} and, optionally, the planes defined in \code{planes_chosen}. The metadata are recompiled based on the new composition of the object.
}
\description{
Subset a \code{segmentation} class object to only contain certain structures
}
