% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldcuts.R
\name{seg_feature_plot}
\alias{seg_feature_plot}
\title{Plot numeric data on a segmentation}
\usage{
seg_feature_plot(
  segmentation,
  feature,
  assay,
  projection = NULL,
  plane = "sagittal",
  smooth = TRUE,
  smoothness = 3,
  min_points = 5,
  color_pal = NULL,
  show_labels = TRUE,
  remove_axes = TRUE
)
}
\arguments{
\item{segmentation}{a \code{segmentation} class object.}

\item{feature}{character, feature to be plotted from the \code{assay} slot.}

\item{assay}{character, the name of the \code{assay} slot in the \code{segmentation}.}

\item{projection}{Character, name of the maximum projection slot to be used? Default is \code{NULL}, which uses the same name as selected in \code{assay}.}

\item{plane}{character, the plane for the maximum projection. Default is "sagittal".}

\item{smooth}{logical, should shapes be smoothed? Default is \code{TRUE}.}

\item{smoothness}{numeric, the smoothing to be used. Default is 3.}

\item{min_points}{numeric, the minimum points to smooth polygons. Default is 5.}

\item{color_pal}{character, the color palette to be used. The default is the `Sunset` palette from \code{colorspace}}

\item{show_labels}{logical, should segmentation labels be shown? Default is \code{TRUE}}

\item{remove_axes}{logical, should axes be shown? Default is \code{FALSE}}
}
\value{
a `ggplot` plot in which structures are coloured according to a numeric value
}
\description{
Creates a `ggplot` plot where numeric data, mapped to structures, fill the segmentation shapes in a colour scale
}
