% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{segmentation-class}
\alias{segmentation-class}
\title{Segmentation class}
\description{
S4 class holding all the data necessary for plotting specific slices, maximum projection of structures, and plotting data from assays
}
\section{Slots}{

\describe{
\item{\code{slices}}{A list containing `segPointSet` instances for the 3 axes ("sagittal", "coronal", "axial")}

\item{\code{ontology}}{A `data.frame` containing the structural ontology}

\item{\code{outlines}}{A list of `data.frame`s containing coordinates for the total outline for each axis}

\item{\code{metadata}}{A list of metadata generated by `drawSegmentation()`}

\item{\code{structure_tables}}{A list of tables with structure per slice per plane generated by `drawSegmentation()`}

\item{\code{assays}}{A list of objects of class \code{segmentationAssay} containing assay data}

\item{\code{projections}}{A list of maximum projections}
}}

