% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{segmentation-class}
\alias{segmentation-class}
\title{S4 class representing a segmentation.}
\description{
Holds all the data necessary for plotting specific slices, maximum projection of structures, recreate the original array.
}
\section{Slots}{

\describe{
\item{\code{slices}}{A list containing `segPointSet` instances for the 3 axes ("sagittal", "coronal", "axial")}

\item{\code{ontology}}{A `data.frame` containing the structural ontology}

\item{\code{outlines}}{A list of `data.frame`s containing coordinates for the total outline for each axis}

\item{\code{structure_tables}}{A matrix containing the number of structures in each slice}

\item{\code{metadata}}{A list of metadata generated by `makeSegmentation()`}
}}

