% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldcuts.R
\name{sliceCheck}
\alias{sliceCheck}
\title{Check slices}
\usage{
sliceCheck(structures, segmentation, planes)
}
\arguments{
\item{structures}{a character containing one or more structure ids}

\item{segmentation}{an object of class \code{segmentation}}

\item{planes}{a vector indicating the planes in which the structure is looked up}
}
\value{
a vector of slices in each plane containing the structure(s) in \code{str}.
}
\description{
Checks which slices in a polygon list contain a given structure id
}
